requirements.txt
pandas>=1.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0

# Machine Learning & Clustering
scikit-learn>=1.0.0

# Visualization
plotly>=5.0.0
streamlit>=1.20.0
streamlit-option-menu>=0.3.0
